+++ src/molpot/pipeline/train/info.md
@@ -1,0 +1,2 @@
+1. Zweryfikować ładowanie się dataset 
+2. Zweryfikowac struktore kodu, poszukac innych przykładow
